Новые решения

+++ Геометрические_задачи/Задача_9.cpp
@@ -1,19 +1,20 @@
 #include <iostream>
 #include <algorithm>
+#include <cmath>
 
-int rotate(int x1, int y1, int x2, int y2, int x3, int y3) {
-    int opr = (x2 - x1)*(y3 - y1) - (y2 - y1)*(x3 - x1);
+int rotate(double x1, double y1, double x2, double y2, double x3, double y3) {
+    double opr = (x2 - x1)*(y3 - y1) - (y2 - y1)*(x3 - x1);
     if (opr > 0) return 1;
     if (opr < 0) return -1;
     return 0;
 }
 
-bool naotrezke(int x1, int y1, int qx, int qy, int x2, int y2) {
+bool naotrezke(double x1, double y1, double qx, double qy, double x2, double y2) {
     return (qx >= std::min(x1,x2) && qx <= std::max(x1,x2)) &&
            (qy >= std::min(y1,y2) && qy <= std::max(y1,y2));
 }
 
-bool perecex(int x1,int y1,int x2,int y2,int x3,int y3,int x4,int y4) {
+bool perecex(double x1,double y1,double x2,double y2,double x3,double y3,double x4,double y4) {
     int o1 = rotate(x1,y1, x2,y2, x3,y3);
     int o2 = rotate(x1,y1, x2,y2, x4,y4);
     int o3 = rotate(x3,y3, x4,y4, x1,y1);
@@ -30,11 +31,11 @@ bool perecex(int x1,int y1,int x2,int y2,int x3,int y3,int x4,int y4) {
 
 bool polygonsIntersect(double poly1[][2], int n1, double poly2[][2], int n2) {
     for (int i = 0; i < n1; ++i) {
-        int x1 = poly1[i][0], y1 = poly1[i][1];
-        int x2 = poly1[(i+1)%n1][0], y2 = poly1[(i+1)%n1][1];
+        double x1 = poly1[i][0], y1 = poly1[i][1];
+        double x2 = poly1[(i+1)%n1][0], y2 = poly1[(i+1)%n1][1];
         for (int j = 0; j < n2; ++j) {
-            int x3 = poly2[j][0], y3 = poly2[j][1];
-            int x4 = poly2[(j+1)%n2][0], y4 = poly2[(j+1)%n2][1];
+            double x3 = poly2[j][0], y3 = poly2[j][1];
+            double x4 = poly2[(j+1)%n2][0], y4 = poly2[(j+1)%n2][1];
             if (perecex(x1,y1,x2,y2,x3,y3,x4,y4)) return true;
         }
     }
